Merge, delete, and rename timeindex tut scripts
The content is almost the sama, except the script to test segmentation
is dropped. We focus on optimisation.
