--- tutorials/advanced/time_index/create_timeseries.py
@@ -1,23 +1,0 @@
-import pandas as pd
-
-
-def reshape_unevenly(df):
-    def to_bucket(ts: pd.Timestamp) -> pd.Timestamp:
-        h = ts.hour
-        d = ts.normalize()
-        if 5 <= h <= 20:
-            return ts
-        if h in (21, 22, 23):
-            return d + pd.Timedelta(hours=21)
-        if h in (0,):
-            return (d - pd.Timedelta(days=1)) + pd.Timedelta(hours=21)
-        # h in (1, 2, 3, 4, 5)
-        return d + pd.Timedelta(hours=1)
-
-    buckets = df.index.map(to_bucket)
-    buckets = buckets.where(buckets >= df.index[0], df.index[0])
-
-    df_mean = df.groupby(buckets).mean().sort_index()
-    df_mean.index.name = "timestamp"
-
-    return df_mean
